re-add ancestor branch detection
This was pulled out because it reportedly did not work very well.  But
the fast-output stream indicates nothing has replaced this feature; all
branches seem to import separately.  Add this funtionality back in until
a better solution is in place.

Note that this feature is enabled unconditionally with this patch,
primarily because the -A option that was used to conditionally enable it
is now used for something else.

Signed-off-by: Chris Rorvick <chris@rorvick.com>